Actually fixed encryption for passwords
After messing with my home system for a bit, I realized that the encryption was not actually working and my password was getting rejected. I found a different decompiler which generated different (more sensible) code which is now generating the same password as the desktop Pentair app and is able to remotely login. This code needs a lot of manual deobfuscation.